Refactor video fetch engine
